feat: mobile view of Split View demo details
